Correct formatting/move files
