+++ src/main/scala/eu/neverblink/jelly/cli/App.scala
@@ -2,7 +2,7 @@ package eu.neverblink.jelly.cli
 
 import caseapp.*
 import eu.neverblink.jelly.cli.command.*
-import eu.neverblink.jelly.cli.command.rdf.RDFFromJelly
+import eu.neverblink.jelly.cli.command.rdf.*
 
 /** Main entrypoint.
   */
@@ -12,5 +12,5 @@ object App extends CommandsEntryPoint:
   override def commands: Seq[Command[?]] = Seq(
     FoolAround,
     Version,
-    RDFFromJelly,
+    RdfFromJelly,
   )
